requirements.txt
absl-py==1.4.0
accelerate==1.0.1
datasets==2.21.0
evaluate==0.4.1
gin-config==0.5.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.10
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.4
pandas==2.0.3
peft==0.11.0
pillow==10.2.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
torch==2.3.0
transformers==4.41.0
wandb==0.15.10